# Optimize DB queries for pause / resume